refactor: added ContactFormWrapper and separated form logic from the presentation layer
